+++ commanderbot/ext/jira/__init__.py
@@ -1,7 +1,8 @@
 from discord.ext.commands import Bot
 
+from commanderbot.core.utils import add_configured_cog
 from commanderbot.ext.jira.jira_cog import JiraCog
 
 
 def setup(bot: Bot):
-    bot.add_cog(JiraCog(bot))
+    add_configured_cog(bot, __name__, JiraCog)